Refactor + JavaDoc
