+++ src/main/java/org/abstractica/javacsg/example/TextExample.java
@@ -2,9 +2,7 @@ package org.abstractica.javacsg.example;
 
 import org.abstractica.javacsg.Geometry2D;
 import org.abstractica.javacsg.JavaCSG;
-import org.abstractica.javacsg.JavaCSGBase;
 import org.abstractica.javacsg.JavaCSGFactory;
-import org.abstractica.javacsg.impl.javaopenscad.JavaCSGBaseOpenSCADImpl;
 
 public class TextExample
 {
